pyproject.toml
[tool.poetry]
name = "tenzor"
version = "0.1.0"
description = ""
authors = ["anton <sparco114@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest-selenium = "^4.0.2"
termcolor = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
